\name{wiDButil}

\alias{wiDButil}

\title{Query And Retrieve Data From The Waterisotopes Database}

\description{
Interface to the public API of the wiDB. Builds, checks and submits queries, and receives and unpacks servers responses.
}

\author{
\strong{Maintainer}: Gabriel Bowen \email{gabe.bowen@utah.edu}  
\strong{Authors}: Gabriel Bowen
}